requirements.txt
# Core ML and AI libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# Image processing
Pillow>=9.5.0
opencv-python>=4.7.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Data handling
pandas>=2.0.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0

# Optional: For GPU acceleration (uncomment if using CUDA)
# torch>=2.0.0+cu118 --index-url https://download.pytorch.org/whl/cu118
# torchaudio>=2.0.0+cu118 --index-url https://download.pytorch.org/whl/cu118

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0